Refactor: Update import statements for custom_logger to use the utils module across multiple files
